:truck: Move Toaster to its single usage place
